move format time into utils.py
